Apply ObservableObject property wrapper. Remove protocol. Fix duplicate file due to rebase merge
